refactor(web): declutter the build viewer header
The header read as four stacked badges plus an action row. Consolidate the
metadata into one quiet line — endo/forma stay as subtle stat pills (the
build's headline numbers) while likes/views and the updated time recede to
muted text — and drop the duplicate like count (it already lives on the heart
button). Fold Share into the overflow (...) menu as Copy link / Copy embed
code, removing the standalone Share button, so the action row is just
like / bookmark / Edit / more. Deletes the now-unused share-menu.tsx.
